Migrate from binary to file_contexts in text format
We previously used a comprehensive file_contexts file in binary format
i.e. file_contexts.bin which contained SELinux labels for all files in
the entire Android system. For our purposes, however, we only need the
few bits that are relevant to building a vendor_dlkm partition image. In
addition, it turned out that the binary format changed somewhat
frequently which led to compatibility issues when this file and the
tools that operated on it got out-of-sync.

The tool is flexible enough to either except a binary file or a text
file as input. Let's take advantage of this and migrate to a text file
that only contains SELinux labels that are relevant to building a
vendor_dlkm partition image.

Bug: 202205395
Signed-off-by: Daniel Mentz <danielmentz@google.com>
Change-Id: I3618541e6c9fbcbd64e9ca2e6433cb4561196348
Signed-off-by: Carlos Llamas <cmllamas@google.com>
(cherry picked from commit 1f6aed53c6cf5b674725854f8f7892569ec0d4a2)